+++ app/Http/Controllers/WisataController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 
 
@@ -19,6 +20,13 @@ class WisataController extends Controller
         $wisata = DB::table('wisata')->get();
         return view('visitor.wisata.index', compact('wisata'));
     }
+
+    public function artikel($id) {
+        $wisata = DB::table('wisata')->where('id_wisata', $id)->first();
+        $pageTitle = 'Pariwisata';
+        
+        return view('visitor.wisata.view', compact('wisata', 'pageTitle'));
+    }    
 
     public function input(Request $request) {
         // Validasi data yang diunggah oleh pengguna
@@ -47,8 +55,12 @@ class WisataController extends Controller
             'kontak' => $validasiData['kontak'],
             'notelp' => $validasiData['notelp'],
             'gambar' => $path,
+            'alamat' => $request->input('alamat'),
+            'link_post_ig' => $request->input('link_post_ig'),
+            'link_post_tiktok' => $request->input('link_post_tiktok'),
+            'link_post_yt' => $request->input('link_post_yt'),
+            'created_by' => $request->input('created_by'),
             'created_at' => now(),
-            'updated_at' => now(),
         ]);
     
         return redirect('/wisata-admin')->with('success', 'Berhasil menambahkan data produk');
@@ -61,32 +73,72 @@ class WisataController extends Controller
     }
     
     public function update(Request $request) {
+        // return $request;
         
         // Ambil data dari form
         $namaWisata = $request->input('wisata');
-        $kontakWisata = $request->input('kontak');
+        $kontakWIsata = $request->input('kontak');
         $deskripsiWisata = $request->input('deskripsi');
         $noTelepon = $request->input('notelp');
-        $gambarWisata = $request->input('gambar');
-    
+        $linkIg = $request->input('link_post_ig');
+        $linkTiktok = $request->input('link_post_tiktok');
+        $linkYt = $request->input('link_post_yt');
+        $alamatWisata = $request->input('alamat');
+
+        $singkatWisata = Str::limit(strip_tags($request->input('deskripsi')), 200);
+
+        $oldImagePath = DB::table('wisata')->where('id_wisata', $request->input('id_wisata'))->value('gambar');
+        
+        // Ambil gambar yang diunggah 
+        if ($request->hasFile('gambar')) {
+            // Hapus gambar lama jika ada
+            if ($oldImagePath) {
+                Storage::delete($oldImagePath);
+            }
+        
+            // Simpan gambar baru
+            $path = $request->file('gambar')->store('gambar-gendro');
+        } else {
+            $path = $oldImagePath;
+        }        
+
         // Update data produk dalam database
         DB::table('wisata')
             ->where('id_wisata', $request->input('id_wisata'))
             ->update([
                 'wisata' => $namaWisata,
-                'kontak' => $kontakWisata,
+                'kontak' => $kontakWIsata,
                 'deskripsi' => $deskripsiWisata,
                 'notelp'=> $noTelepon,
-                'gambar'=> $gambarWisata,
+                'link_post_ig' => $linkIg,
+                'link_post_tiktok' => $linkTiktok,
+                'link_post_yt' => $linkYt,
+                'gambar'=> $path,
+                'alamat' => $alamatWisata,
+                'updated_at' => now(),
+                'singkat' => $singkatWisata,
             ]);
     
         // Redirect atau kirim respons sesuai kebutuhan Anda
-        return response()->json(['url' => '/wisata-admin', 'message' => 'Berhasil mengubah data', 'success' => true]);    
+        return redirect('/wisata-admin')->with('success', 'Berhasil mengubah data wisata.');
     }
 
     public function delete($id)
     {
-        DB::table('wisata')->where('id_wisata', $id)->delete();
-        return redirect('/wisata-admin')->with('success', 'Berhasil hapus wisata.');
+        $wisata = DB::table('wisata')->where('id_wisata', $id)->first();
+    
+        if ($wisata) {
+    
+            if (!empty($wisata->gambar)) {
+                Storage::delete($wisata->gambar);
+            }
+    
+            // Hapus entri wisata dari database
+            DB::table('wisata')->where('id_wisata', $id)->delete();
+            
+            return redirect('/wisata-admin')->with('success', 'Berhasil hapus tempat wisata.');
+        } else {
+            return redirect('/wisata-admin')->with('error', 'Tempat wisata tidak ditemukan.');
+        }
     }
 }
